requirements: Create separate list for testing
Earlier we had a mix of packages needed for testing and for
installing in the requirements.txt. But, packages like coverage
are not needed to be installed before using coala. Hence,
it makes sense to separate the packages required for running and
those needed for testing coala.
This commit adds the test-requirements.txt file which does this.

Fixes https://github.com/coala-analyzer/coala/issues/1293