Tests: Remove sys.path manipulation for run_tests
run_tests.py required us to modify the sys.path in every test. Now
that we are using pytest, this is not needed as it handles testing
in a different way. Hence, we can remove those statements.

`git grep sys.path` was used to make sure no instance was left
out.

Related to https://github.com/coala-analyzer/coala/issues/868